add auth with Clerk
