+++ .env
@@ -1,1 +1,5 @@
 OPENAI_API_KEY=dummy
+NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY=dummy
+CLERK_SECRET_KEY=dummy
+NEXT_PUBLIC_CLERK_AFTER_SIGN_IN_URL=/
+NEXT_PUBLIC_CLERK_AFTER_SIGN_UP_URL=/
