+++ package.json
@@ -9,6 +9,8 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@clerk/nextjs": "^4.29.12",
+    "@clerk/themes": "^1.7.12",
     "@tailwindcss/typography": "^0.5.10",
     "@types/formidable": "^3.4.5",
     "@types/node": "20.11.20",
@@ -17,14 +19,14 @@
     "@types/react-dom": "18.2.7",
     "@types/react-syntax-highlighter": "^15.5.11",
     "@vercel/analytics": "^1.2.2",
-    "ai": "^2.2.31",
+    "ai": "^3.0.19",
     "autoprefixer": "^10.4.17",
-    "eslint": "8.47.0",
-    "eslint-config-next": "13.4.19",
+    "eslint": "^9.0.0",
+    "eslint-config-next": "^14.1.4",
     "formidable": "^3.5.1",
     "fs": "^0.0.1-security",
-    "next": "13.4.19",
-    "openai": "^4.24.2",
+    "next": "^14.1.4",
+    "openai": "^4.33.0",
     "pdfjs-dist": "^4.0.379",
     "postcss": "^8.4.35",
     "puppeteer": "^22.2.0",
@@ -37,6 +39,6 @@
     "remark-gfm": "^3.0.1",
     "remark-math": "^5.1.1",
     "tailwindcss": "^3.4.1",
-    "typescript": "5.3.3"
+    "typescript": "5.4.4"
   }
 }
